Fix to not ask if you want to submit when you run out of time
The timed exam was asking if you wanted to submit when you had run out
of time.  I changed it to tell you that you ran out of time and your
answers have been saved.  It still asks if you want to submit if you
clicked the button to finish the exam.