Add section to contents of activecode to contain the question.
The section for adding explanatory text to an activecode that we intend to use as an exercise, starts at the beginning of the body and goes to a line that contains ~~~~  (four tilde)  This section is recursively parsed by Sphinx and so may contain any sphinx markup you wish.  The line with the ~ is discarded and the remainder of the body is processed as usual.