Fix: Score multiple choice, multiple-answer questions correctly.
If a students answers an MCMA question with all the correct answers and one or more incorrect answers, the existing code sets this.correct = true (so that the students get credit for getting the problem correct in the gradebook) and also shows that the problem is incorrect (so that the student thinks their answer is wrong, as it indeed is). If a student corrects their answer, it won't be saved (ajax.py doesn't overwrite correct answers) -- refreshing will again display the "incorrect" answer. (This change doesn't correct server-side bugs, just prevents it from happening in the future).