Fix warnings and add python_requires >= 3.6
also contains experimental code to do pre-install and post-install
hooks for running scripts.  This would be very useful for times when
we need to require some non-python stuff.
We may even be able to use this to check for npm and then post development
install (-e) run npm run build.