A bit of proactive defense
There are times when the first four lines of a unit test for a Java program are missing, causing the tests to fail.  I do not know where, when, or why this is happening.  But this checks to see if the important things are missing right before they are sent to JOBE. If they are missing I add them.